requirements.txt
cffi==1.14.5
click==7.1.2
Pillow==8.1.2
pycparser==2.20
pyheif==0.5.1